refactor: colocalize nodesecurerc.ts with /config domain
