+++ src/reporters/internal/scanner.ts
@@ -3,8 +3,8 @@ import Spinner from "@slimio/async-cli-spinner";
 import pluralize from "pluralize";
 import ms from "pretty-ms";
 
+import * as RC from "../../config/nodesecurerc.js";
 import { consolePrinter } from "../../lib/console-printer/index.js";
-import * as RC from "../../nodesecurerc.js";
 import { Reporter } from "../reporter.js";
 
 /**
